Moved al-atik into utils_gmpes for improved efficiency of code (best to minimise calls from sm_utils based on previous documentation by Graeme)
