+++ openquake/smt/residuals/gmpe_residuals.py
@@ -41,7 +41,8 @@ from openquake.hazardlib.gsim import get_available_gsims
 from openquake.hazardlib import imt
 import openquake.smt.intensity_measures as ims
 from openquake.smt.strong_motion_selector import SMRecordSelector
-from openquake.smt.sm_utils import convert_accel_units, check_gsim_list, al_atik_sigma_check
+from openquake.smt.sm_utils import convert_accel_units, check_gsim_list
+from openquake.smt.comparison.utils_gmpes import  al_atik_sigma_check
 
 GSIM_LIST = get_available_gsims()
 GSIM_KEYS = set(GSIM_LIST)
